ingesting/requirements.txt
# Web API
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.3.0
httpx==0.28.1
python-dotenv==1.1.1   # nếu bạn load .env
Deprecated==1.2.18

# GCS
google-cloud-storage==3.2.0

# Pinecone v3
pinecone==7.3.0

# Observability
loguru==0.7.3
prometheus_client==0.22.1
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-prometheus==0.57b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-semantic-conventions==0.57b0
